Fixing bugs

+++ src/main/java/com/entropyteam/entropay/employees/dtos/OvertimeDto.java
@@ -10,17 +10,20 @@ import java.util.UUID;
 import jakarta.validation.constraints.NotNull;
 
 public record OvertimeDto(UUID id,
-                          int hours,
+                          float hours,
                           String description,
-                          @JsonFormat(pattern = "yyyy-MM-dd") LocalDate date,
+                          @JsonFormat(pattern = "yyyy-MM-dd")
+                          LocalDate date,
                           @NotNull(message ="EmployeeId is mandatory")
                           UUID employeeId,
                           @NotNull(message ="AssignmentId is mandatory")
                           UUID assignmentId,
                           boolean deleted,
-                          @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime createdAt,
-                          @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime modifiedAt
-                          )  {
+                          @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+                          LocalDateTime createdAt,
+                          @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+                          LocalDateTime modifiedAt)  {
+
     public OvertimeDto (Overtime overtime){
         this(overtime.getId(), overtime.getHours(), overtime.getDescription(), overtime.getDate(),
                 overtime.getEmployee().getId(), overtime.getAssignment().getId(),
