+++ src/main/resources/db/migration/V61.0__DDL_create_overtime_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE overtime (
     modified_at  TIMESTAMP 					 NOT NULL,
     deleted      BOOLEAN                     NOT NULL,
     date   TIMESTAMP 					 NOT NULL,
-    hours INTEGER,
+    hours FLOAT,
     description VARCHAR(100),
     employee_id  UUID,
     assignment_id     UUID,
